refactor: extract the durable federated-media executor out of the audio adapter
The retry, idempotent-replay, cancellation, and hash-verified-transfer logic in
audioGen/remote.js is the part that has to behave identically for every media
kind, and image/video adapters were about to copy it twice. Move it to
services/federatedMedia/remoteExecutor.js, parameterized by kind, marker shape,
wire-request builder, result destination, and a local-registration step.

Behavior-preserving: audioGen/remote.js keeps its exports and its existing suite
passes unchanged. One detail is now pinned by a comment rather than by accident
— the audio submission still sends no explicit `kind`, because an already
deployed audio-only provider validates the body against a strict schema that
predates that field and would reject it.

Refs #4348
